requirements.txt
Django==4.2.14
python-decouple==3.8
gunicorn==23.0.0
whitenoise==6.8.2
django-csp==4.0
reportlab==4.4.9
openpyxl==3.1.5
Pillow==12.1.0
asgiref==3.11.0
sqlparse==0.5.5
psycopg2-binary==2.9.10
dj-database-url==2.3.0
